refactor: extract parent→child constants and break DraftMatecatUtils bilateral cycles
Break 15 circular dependency cycles (40→25) via:
- OnBoarding: extract context/socialUrls to OnBoardingContext.js (5 cycles)
- ActivityLog: extract context/constants to sibling files (3 cycles)
- SegmentFooterTabGlossary: extract icons/constants to GlossaryConstants.js (2 cycles)
- UploadFile: extract getPrintableFileSize to UploadFileUtils.js (2 cycles)
- SettingsPanel: move DEFAULT_ENGINE_MEMORY to SettingsPanelConstants.js (1 cycle)
- DraftMatecatUtils: extract regexWordDelimiter to textConstants.js (3 cycles)

Cumulative: 132→25 (81% reduction). Plugins verified clean.
