+++ public/js/components/activityLog/ActivityLogConstants.js
@@ -1,0 +1,10 @@
+export const ACTIVITY_LOG_COLUMNS = [
+  {id: 'ip', label: 'User IP'},
+  {id: 'event_date', label: 'Event Date', sortingType: 'date'},
+  {id: 'id_project', label: 'Project ID'},
+  {id: 'id_job', label: 'Job ID'},
+  {id: 'languagePair', label: 'Language Pair'},
+  {id: 'userName', label: 'User Name'},
+  {id: 'email', label: 'User Email'},
+  {id: 'action', label: 'Action'},
+]
