+++ public/js/components/activityLog/ActivityLogContext.js
@@ -1,0 +1,3 @@
+import {createContext} from 'react'
+
+export const ActivityLogContext = createContext({})
