+++ public/js/components/activityLog/ActivityLogTable.js
@@ -1,18 +1,10 @@
 import React, {useContext, useState} from 'react'
 import {ColumnSorting} from './ColumnSorting'
-import {ActivityLogContext} from '../../pages/ActivityLog'
+import {ActivityLogContext} from './ActivityLogContext'
 import LabelWithTooltip from '../common/LabelWithTooltip'
 
-export const ACTIVITY_LOG_COLUMNS = [
-  {id: 'ip', label: 'User IP'},
-  {id: 'event_date', label: 'Event Date', sortingType: 'date'},
-  {id: 'id_project', label: 'Project ID'},
-  {id: 'id_job', label: 'Job ID'},
-  {id: 'languagePair', label: 'Language Pair'},
-  {id: 'userName', label: 'User Name'},
-  {id: 'email', label: 'User Email'},
-  {id: 'action', label: 'Action'},
-]
+import {ACTIVITY_LOG_COLUMNS} from './ActivityLogConstants'
+export {ACTIVITY_LOG_COLUMNS}
 
 export const ActivityLogTable = () => {
   const {activityLog, filterByColumn} = useContext(ActivityLogContext)
