+++ public/js/components/createProject/UploadFileLocal.js
@@ -1,12 +1,12 @@
 import React, {useContext, useMemo} from 'react'
 import {Button, BUTTON_SIZE, BUTTON_TYPE} from '../common/Button/Button'
-import {DeleteIcon} from '../segments/SegmentFooterTabGlossary'
+import {DeleteIcon} from '../segments/SegmentFooterTabGlossary/GlossaryConstants'
 import FileUploadIconBig from '../../../img/icons/FileUploadIconBig'
 import CommonUtils from '../../utils/commonUtils'
 import IconAdd from '../icons/IconAdd'
 import IconClose from '../icons/IconClose'
 import {PROGRESS_BAR_SIZE, ProgressBar} from '../common/ProgressBar'
-import {getPrintableFileSize} from './UploadFile'
+import {getPrintableFileSize} from './UploadFileUtils'
 import {CreateProjectContext} from './CreateProjectContext'
 import {useFileUploadManager} from './hooks/useFileUploadManager'
 import {useDragAndDrop} from './hooks/useDragAndDrop'
