+++ public/js/components/createProject/UploadFileUtils.js
@@ -1,0 +1,9 @@
+export const getPrintableFileSize = (filesizeInBytes) => {
+  filesizeInBytes = filesizeInBytes / 1024
+  let ext = ' KB'
+  if (filesizeInBytes > 1024) {
+    filesizeInBytes = filesizeInBytes / 1024
+    ext = ' MB'
+  }
+  return Math.round(filesizeInBytes * 100, 2) / 100 + ext
+}
