+++ public/js/components/createProject/UploadGdrive.js
@@ -1,8 +1,8 @@
 import React, {useContext, useEffect, useMemo} from 'react'
 import CommonUtils from '../../utils/commonUtils'
-import {getPrintableFileSize} from './UploadFile'
+import {getPrintableFileSize} from './UploadFileUtils'
 import {Button, BUTTON_SIZE, BUTTON_TYPE} from '../common/Button/Button'
-import {DeleteIcon} from '../segments/SegmentFooterTabGlossary'
+import {DeleteIcon} from '../segments/SegmentFooterTabGlossary/GlossaryConstants'
 import IconClose from '../icons/IconClose'
 import {usePrevious} from '../../hooks/usePrevious'
 import {CreateProjectContext} from './CreateProjectContext'
