+++ public/js/components/segments/utils/DraftMatecatUtils/textConstants.js
@@ -1,0 +1,2 @@
+export const regexWordDelimiter =
+  /(\s+|[-+*\\/]|\d+|[!@#$%^&*()_+\-=\\[\]{};':"\\|,.<>\\/?~°⇥])/
